add import

+++ frameworks/taihe/media/src/avtranscoder_taihe.cpp
@@ -179,14 +179,14 @@ RetInfo AVTranscoderImpl::SetOutputFile(int32_t fd)
     return RetInfo(MSERR_EXT_API9_OK, "");
 }
 
-double AVTranscoderImpl::GetFdDst()
+int32_t AVTranscoderImpl::GetFdDst()
 {
     MediaTrace trace("AVTranscoderImpl::get url");
     MEDIA_LOGD("TaiheGetUrl Out Current Url: %{public}s", srcUrl_.c_str());
     return dstFd_;
 }
 
-void AVTranscoderImpl::SetFdDst(double fdDst)
+void AVTranscoderImpl::SetFdDst(int32_t fdDst)
 {
     MediaTrace trace("AVTranscoderImpl::set fd");
     auto asyncCtx = std::make_unique<AVTransCoderAsyncContext>();
@@ -215,8 +215,8 @@ ohos::multimedia::media::AVFileDescriptor AVTranscoderImpl::GetFdSrc()
     MediaTrace trace("AVTranscoderImpl::get url");
     ohos::multimedia::media::AVFileDescriptor fdSrc;
     fdSrc.fd = srcFd_.fd;
-    fdSrc.offset = optional<double>(std::in_place_t{}, srcFd_.offset);
-    fdSrc.length = optional<double>(std::in_place_t{}, srcFd_.length);
+    fdSrc.offset = optional<int64_t>(std::in_place_t{}, srcFd_.offset);
+    fdSrc.length = optional<int64_t>(std::in_place_t{}, srcFd_.length);
     MEDIA_LOGD("TaiheGetUrl Out Current Url: %{public}s", srcUrl_.c_str());
     return fdSrc;
 }
@@ -510,15 +510,15 @@ void AVTranscoderImpl::OffComplete(optional_view<callback<void(uintptr_t)>> call
     MEDIA_LOGI("OffComplete End");
 }
 
-void AVTranscoderImpl::OnProgressUpdate(callback_view<void(double)> callback)
+void AVTranscoderImpl::OnProgressUpdate(callback_view<void(int32_t)> callback)
 {
     MediaTrace trace("AVTranscoderImpl::OnProgressUpdate");
     MEDIA_LOGI("OnProgressUpdate Start");
 
     std::string callbackName = STATE_PROGRESSUPDATE;
     ani_env *env = get_env();
-    std::shared_ptr<taihe::callback<void(double)>> taiheCallback =
-            std::make_shared<taihe::callback<void(double)>>(callback);
+    std::shared_ptr<taihe::callback<void(int32_t)>> taiheCallback =
+            std::make_shared<taihe::callback<void(int32_t)>>(callback);
     std::shared_ptr<uintptr_t> cacheCallback = std::reinterpret_pointer_cast<uintptr_t>(taiheCallback);
 
     std::shared_ptr<AutoRef> autoRef = std::make_shared<AutoRef>(env, cacheCallback);
@@ -526,7 +526,7 @@ void AVTranscoderImpl::OnProgressUpdate(callback_view<void(double)> callback)
     MEDIA_LOGI("OnProgressUpdate End");
 }
 
-void AVTranscoderImpl::OffProgressUpdate(optional_view<callback<void(double)>> callback)
+void AVTranscoderImpl::OffProgressUpdate(optional_view<callback<void(int32_t)>> callback)
 {
     MediaTrace trace("AVTranscoderImpl::OffProgressUpdate");
     MEDIA_LOGI("OffProgressUpdate Start");
